requirements.txt
pypdf==5.4.0
sentence-transformers==3.4.1
faiss-cpu==1.12.0
numpy==2.2.4
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3

